Create docking_functions module for functions relating to docking ;-)
Also added include into robot_defines.h
