+++ RobotNoASF/docking_functions.h
@@ -1,0 +1,45 @@
+/*
+* docking_functions.h
+*
+* Author : Matthew Witt (pxf5695@autuni.ac.nz)
+* Created: 25/07/2017 8:08:11 PM
+*
+* Project Repository: https://github.com/AdamParlane/aut-swarm-robotics
+*
+* [WIP] Contains functions for docking... [WIP]
+*
+* More Info:
+* Atmel SAM 4N Processor Datasheet:http://www.atmel.com/Images/Atmel-11158-32-bit%20Cortex-M4-Microcontroller-SAM4N16-SAM4N8_Datasheet.pdf
+* Relevant reference materials or datasheets if applicable
+*
+* Functions:
+* void dockRobot(void)
+*
+*/
+
+#ifndef DOCKING_FUNCTIONS_H_
+#define DOCKING_FUNCTIONS_H_
+
+///////////////Includes/////////////////////////////////////////////////////////////////////////////
+#include "robot_defines.h"
+
+///////////////Defines//////////////////////////////////////////////////////////////////////////////
+
+
+///////////////Functions////////////////////////////////////////////////////////////////////////////
+/*
+* Function:
+* void dockRobot(void)
+*
+* Function to guide the robot to the dock.
+*
+* Inputs:
+* none
+*
+* Returns:
+* none
+*
+*/
+void dockRobot(void);
+
+#endif /* DOCKING_FUNCTIONS_H_ */
